# [CON-71] Parallelize aggregateReconfigAndPotentialSyncOps in chunks of 500 users