… creator-node/test/snapbackSM.test.js