### Description

- Change loop of awaits to Promise.allSettled in batches of 500
- Clean up old logic for supporting outdated nodes that don't have the spID fields for primary and secondaries

<details>
  <summary>
    Logs from testing on CN10 (~9x faster for ~550 users)
  </summary>

#### Without Parallelization
  <details>
    <summary>
      First Run (569 users in 23.22 minutes)
    </summary>

```
[{"stage":"BEGIN processStateMachineOperation()","data":{"currentModuloSlice":22,"moduloBase":50},"time":1651486620982},{"stage":"getNodeUsers() and sliceUsers() Success","data":{"nodeUsersLength":569},"time":1651486627659,"duration":6677},{"stage":"getUnhealthyPeers() Success","data":{"unhealthyPeerSetLength":1,"unhealthyPeers":["https://creatornode7.staging.audius.co"]},"time":1651486628554,"duration":895},{"stage":"buildReplicaSetNodesToUserWalletsMap() Success","data":{"numReplicaSetNodes":8},"time":1651486628555,"duration":1},{"stage":"updateEndpointToSpIdMap() Success","data":{"endpointToSPIdMapSize":8},"time":1651486628922,"duration":367},{"stage":"retrieveClockStatusesForUsersAcrossReplicaSet() Success","data":{},"time":1651486629424,"duration":502},{"stage":"Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays","data":{"requiredUpdateReplicaSetOpsLength":200,"potentialSyncRequestsLength":938},"time":1651488022538,"duration":1393114},{"stage":"issueSyncRequestsToSecondaries() Success","data":{"numSyncRequestsRequired":2,"numSyncRequestsEnqueued":2,"numEnqueueSyncRequestErrors":0,"enqueueSyncRequestErrors":[]},"time":1651488022543,"duration":5},{"stage":"issueUpdateReplicaSetOp() Error","data":{"error":"issueUpdateReplicaSetOp() failed for 19 users"},"time":1651489792913,"duration":1770370},{"stage":"processStateMachineOperation() Error","data":{"error":"processStateMachineOperation():issueUpdateReplicaSetOp() Error"},"time":1651489792913,"duration":0},{"stage":"END processStateMachineOperation()","data":{"currentModuloSlice":22,"moduloBase":50},"time":1651489792913,"duration":0,"fullDuration":3171931}]
```

  </details>

  <details>
      <summary>
        Second Run (563 users in 24.38 minutes)
      </summary>
  
```
[{"stage":"BEGIN processStateMachineOperation()","data":{"currentModuloSlice":23,"moduloBase":50},"time":1651489792920},{"stage":"getNodeUsers() and sliceUsers() Success","data":{"nodeUsersLength":563},"time":1651489799521,"duration":6601},{"stage":"getUnhealthyPeers() Success","data":{"unhealthyPeerSetLength":1,"unhealthyPeers":["https://creatornode7.staging.audius.co"]},"time":1651489800472,"duration":951},{"stage":"buildReplicaSetNodesToUserWalletsMap() Success","data":{"numReplicaSetNodes":8},"time":1651489800474,"duration":2},{"stage":"updateEndpointToSpIdMap() Success","data":{"endpointToSPIdMapSize":8},"time":1651489800844,"duration":370},{"stage":"retrieveClockStatusesForUsersAcrossReplicaSet() Success","data":{},"time":1651489801341,"duration":497},{"stage":"Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays","data":{"requiredUpdateReplicaSetOpsLength":100,"potentialSyncRequestsLength":1026},"time":1651491264458,"duration":1463117},{"stage":"issueSyncRequestsToSecondaries() Success","data":{"numSyncRequestsRequired":2,"numSyncRequestsEnqueued":2,"numEnqueueSyncRequestErrors":0,"enqueueSyncRequestErrors":[]},"time":1651491264463,"duration":5},{"stage":"issueUpdateReplicaSetOp() Error","data":{"error":"issueUpdateReplicaSetOp() failed for 11 users"},"time":1651492106697,"duration":842234},{"stage":"processStateMachineOperation() Error","data":{"error":"processStateMachineOperation():issueUpdateReplicaSetOp() Error"},"time":1651492106697,"duration":0},{"stage":"END processStateMachineOperation()","data":{"currentModuloSlice":23,"moduloBase":50},"time":1651492106697,"duration":0,"fullDuration":2313777}]
```
  
  </details>

#### With Parallelization
  <details>
    <summary>
      First Run (568 users in 2.83 minutes)
    </summary>

```
[{"stage":"BEGIN processStateMachineOperation()","data":{"currentModuloSlice":24,"moduloBase":50},"time":1651492106702},{"stage":"getNodeUsers() and sliceUsers() Success","data":{"nodeUsersLength":1509},"time":1651492112199,"duration":5497},{"stage":"getUnhealthyPeers() Success","data":{"unhealthyPeerSetLength":1,"unhealthyPeers":["https://creatornode7.staging.audius.co"]},"time":1651492113514,"duration":1315},{"stage":"buildReplicaSetNodesToUserWalletsMap() Success","data":{"numReplicaSetNodes":8},"time":1651492113518,"duration":4},{"stage":"updateEndpointToSpIdMap() Success","data":{"endpointToSPIdMapSize":8},"time":1651492113926,"duration":408},{"stage":"retrieveClockStatusesForUsersAcrossReplicaSet() Success","data":{},"time":1651492114591,"duration":665},{"stage":"Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays","data":{"requiredUpdateReplicaSetOpsLength":294,"potentialSyncRequestsLength":989},"time":1651493674005,"duration":1559414},{"stage":"issueSyncRequestsToSecondaries() Success","data":{"numSyncRequestsRequired":2,"numSyncRequestsEnqueued":2,"numEnqueueSyncRequestErrors":0,"enqueueSyncRequestErrors":[]},"time":1651493674011,"duration":6},{"stage":"issueUpdateReplicaSetOp() Error","data":{"error":"issueUpdateReplicaSetOp() failed for 42 users"},"time":1651496280845,"duration":2606834},{"stage":"processStateMachineOperation() Error","data":{"error":"processStateMachineOperation():issueUpdateReplicaSetOp() Error"},"time":1651496280845,"duration":0},{"stage":"END processStateMachineOperation()","data":{"currentModuloSlice":24,"moduloBase":50},"time":1651496280845,"duration":0,"fullDuration":4174143}]
```

  </details>

<details>
        <summary>
          Second Run (573 users in 2.82 minutes)
        </summary>

```
[{"stage":"BEGIN processStateMachineOperation()","data":{"currentModuloSlice":15,"moduloBase":50},"time":1651501388143},{"stage":"getNodeUsers() and sliceUsers() Success","data":{"nodeUsersLength":573},"time":1651501394471,"duration":6328},{"stage":"getUnhealthyPeers() Success","data":{"unhealthyPeerSetLength":1,"unhealthyPeers":["https://creatornode7.staging.audius.co"]},"time":1651501396028,"duration":1557},{"stage":"buildReplicaSetNodesToUserWalletsMap() Success","data":{"numReplicaSetNodes":8},"time":1651501396029,"duration":1},{"stage":"updateEndpointToSpIdMap() Success","data":{"endpointToSPIdMapSize":8},"time":1651501396471,"duration":442},{"stage":"retrieveClockStatusesForUsersAcrossReplicaSet() Success","data":{},"time":1651501397401,"duration":930},{"stage":"Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays","data":{"requiredUpdateReplicaSetOpsLength":12,"potentialSyncRequestsLength":1134},"time":1651501566466,"duration":169065},{"stage":"issueSyncRequestsToSecondaries() Success","data":{"numSyncRequestsRequired":0,"numSyncRequestsEnqueued":0,"numEnqueueSyncRequestErrors":0,"enqueueSyncRequestErrors":[]},"time":1651501566470,"duration":4},{"stage":"issueUpdateReplicaSetOp() Error","data":{"error":"issueUpdateReplicaSetOp() failed for 12 users"},"time":1651501830807,"duration":264337},{"stage":"processStateMachineOperation() Error","data":{"error":"processStateMachineOperation():issueUpdateReplicaSetOp() Error"},"time":1651501830808,"duration":1},{"stage":"END processStateMachineOperation()","data":{"currentModuloSlice":15,"moduloBase":50},"time":1651501830808,"duration":0,"fullDuration":442665}]
```

  </details>

</details>

---

<details>
  <summary>
    Logs from testing on CN11 (~10x faster for ~350 users)
  </summary>

#### Without Parallelization
  <details>
    <summary>
      First Run (365 users in 427.86 seconds)
    </summary>

```
[{"stage":"BEGIN processStateMachineOperation()","data":{"currentModuloSlice":14,"moduloBase":50},"time":1651177257986},{"stage":"getNodeUsers() and sliceUsers() Success","data":{"nodeUsersLength":365},"time":1651177266112,"duration":8126},{"stage":"getUnhealthyPeers() Success","data":{"unhealthyPeerSetLength":0,"unhealthyPeers":[]},"time":1651177267392,"duration":1280},{"stage":"buildReplicaSetNodesToUserWalletsMap() Success","data":{"numReplicaSetNodes":8},"time":1651177267393,"duration":1},{"stage":"updateEndpointToSpIdMap() Success","data":{"endpointToSPIdMapSize":8},"time":1651177267774,"duration":381},{"stage":"retrieveClockStatusesForUsersAcrossReplicaSet() Success","data":{},"time":1651177268484,"duration":710},{"stage":"Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays","data":{"requiredUpdateReplicaSetOpsLength":0,"potentialSyncRequestsLength":730},"time":1651177696342,"duration":427858},{"stage":"issueSyncRequestsToSecondaries() Success","data":{"numSyncRequestsRequired":0,"numSyncRequestsEnqueued":0,"numEnqueueSyncRequestErrors":0,"enqueueSyncRequestErrors":[]},"time":1651177696345,"duration":3},{"stage":"issueUpdateReplicaSetOp() Success","data":{"numUpdateReplicaOpsIssued":0},"time":1651177697119,"duration":774},{"stage":"END processStateMachineOperation()","data":{"currentModuloSlice":14,"moduloBase":50},"time":1651177697119,"duration":0,"fullDuration":439133}]
```

  </details>

  <details>
      <summary>
        Second Run (388 users in 456.74 seconds)
      </summary>

```
[{"stage":"BEGIN processStateMachineOperation()","data":{"currentModuloSlice":42,"moduloBase":50},"time":1651155000009},{"stage":"getNodeUsers() and sliceUsers() Success","data":{"nodeUsersLength":388},"time":1651155007692,"duration":7683},{"stage":"getUnhealthyPeers() Success","data":{"unhealthyPeerSetLength":0,"unhealthyPeers":[]},"time":1651155008846,"duration":1154},{"stage":"buildReplicaSetNodesToUserWalletsMap() Success","data":{"numReplicaSetNodes":8},"time":1651155008847,"duration":1},{"stage":"updateEndpointToSpIdMap() Success","data":{"endpointToSPIdMapSize":8},"time":1651155009256,"duration":409},{"stage":"retrieveClockStatusesForUsersAcrossReplicaSet() Success","data":{},"time":1651155009896,"duration":640},{"stage":"Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays","data":{"requiredUpdateReplicaSetOpsLength":0,"potentialSyncRequestsLength":776},"time":1651155466637,"duration":456741},{"stage":"issueSyncRequestsToSecondaries() Success","data":{"numSyncRequestsRequired":2,"numSyncRequestsEnqueued":2,"numEnqueueSyncRequestErrors":0,"enqueueSyncRequestErrors":[]},"time":1651155466642,"duration":5},{"stage":"issueUpdateReplicaSetOp() Success","data":{"numUpdateReplicaOpsIssued":0},"time":1651155467347,"duration":705},{"stage":"END processStateMachineOperation()","data":{"currentModuloSlice":42,"moduloBase":50},"time":1651155467347,"duration":0,"fullDuration":467338}]
```

  </details>

  #### With Parallelization
  <details>
    <summary>
      First Run (310 users in 39.59 seconds)
    </summary>

```
[{"stage":"BEGIN processStateMachineOperation()","data":{"currentModuloSlice":25,"moduloBase":50},"time":1651273812358},{"stage":"getNodeUsers() and sliceUsers() Success","data":{"nodeUsersLength":310},"time":1651273818453,"duration":6095},{"stage":"getUnhealthyPeers() Success","data":{"unhealthyPeerSetLength":0,"unhealthyPeers":[]},"time":1651273819959,"duration":1506},{"stage":"buildReplicaSetNodesToUserWalletsMap() Success","data":{"numReplicaSetNodes":8},"time":1651273819960,"duration":1},{"stage":"updateEndpointToSpIdMap() Success","data":{"endpointToSPIdMapSize":8},"time":1651273820381,"duration":421},{"stage":"retrieveClockStatusesForUsersAcrossReplicaSet() Success","data":{},"time":1651273821136,"duration":755},{"stage":"Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays","data":{"requiredUpdateReplicaSetOpsLength":0,"potentialSyncRequestsLength":620},"time":1651273860734,"duration":39598},{"stage":"issueSyncRequestsToSecondaries() Success","data":{"numSyncRequestsRequired":0,"numSyncRequestsEnqueued":0,"numEnqueueSyncRequestErrors":0,"enqueueSyncRequestErrors":[]},"time":1651273860736,"duration":2},{"stage":"issueUpdateReplicaSetOp() Success","data":{"numUpdateReplicaOpsIssued":0},"time":1651273861490,"duration":754},{"stage":"END processStateMachineOperation()","data":{"currentModuloSlice":25,"moduloBase":50},"time":1651273861490,"duration":0,"fullDuration":49132}]
```

  </details>

  <details>
      <summary>
        Second Run (329 users in 42.26 seconds)
      </summary>

```
[{"stage":"BEGIN processStateMachineOperation()","data":{"currentModuloSlice":26,"moduloBase":50},"time":1651274741692},{"stage":"getNodeUsers() and sliceUsers() Success","data":{"nodeUsersLength":329},"time":1651274747837,"duration":6145},{"stage":"getUnhealthyPeers() Success","data":{"unhealthyPeerSetLength":0,"unhealthyPeers":[]},"time":1651274749224,"duration":1387},{"stage":"buildReplicaSetNodesToUserWalletsMap() Success","data":{"numReplicaSetNodes":8},"time":1651274749225,"duration":1},{"stage":"updateEndpointToSpIdMap() Success","data":{"endpointToSPIdMapSize":8},"time":1651274749602,"duration":377},{"stage":"retrieveClockStatusesForUsersAcrossReplicaSet() Success","data":{},"time":1651274750064,"duration":462},{"stage":"Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays","data":{"requiredUpdateReplicaSetOpsLength":0,"potentialSyncRequestsLength":658},"time":1651274792323,"duration":42259},{"stage":"issueSyncRequestsToSecondaries() Success","data":{"numSyncRequestsRequired":0,"numSyncRequestsEnqueued":0,"numEnqueueSyncRequestErrors":0,"enqueueSyncRequestErrors":[]},"time":1651274792325,"duration":2},{"stage":"issueUpdateReplicaSetOp() Success","data":{"numUpdateReplicaOpsIssued":0},"time":1651274793041,"duration":716},{"stage":"END processStateMachineOperation()","data":{"currentModuloSlice":26,"moduloBase":50},"time":1651274793041,"duration":0,"fullDuration":51349}]
```

  </details>

</details>


### Tests
1. Comment out the auto-queuing in snapback init
2. Add a route that queues a snapback job. Here's mine:
```js
app.get(
    '/theo/enqueue',
    handleResponse(async (req, res, next) => {
      const serviceRegistry = app.get('serviceRegistry')
      await serviceRegistry.snapbackSM.stateMachineQueue.add(
        /** data */ { startTime: Date.now() },
        /** opts */ {}
      )

      return successResponse()
    })
  )
```
3. Hit the route
4. Search the Docker logs for the duration on the `Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays` step and see if it's shorter than the old way. You can get the time for the old way by reverting this PR's changes and re-running the steps above
    - `docker container ps | grep "cn"`
    - `docker container logs <container for the CN you're looking for> | less`
    - Type `/Build requiredUpdateReplicaSetOps and potentialSyncRequests arrays` <Enter>

### How will this change be monitored? Are there sufficient logs?
- Look for increase in logged error: `processStateMachineOperation() Error` including any new error messages
- Look for newly-created error: `aggregateReconfigAndPotentialSyncOps() encountered unexpected failure`
- Keep an eye out for any errors relating to too many connections ("Max connections exceeded")