@theoilie 2022-04-28
Requesting Sid for general review and Vicky because it looks like you last touched the part for outdated nodes without spID fields 🙂