@theoilie 2022-05-02
Updated description with logs for runtime comparisons on staging (about 8-10x faster). Ready for re-review!